op-acceptance-tests: move all ext hfs tests to a single pkg
