--- op-acceptance-tests/tests/sync_tester/sync_tester_hfs_ext_canyon/sync_tester_hfs_ext_canyon_test.go
@@ -1,17 +1,0 @@
-package sync_tester_hfs_ext_canyon
-
-import (
-	"testing"
-
-	"github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester/hardforks_ext"
-	"github.com/ethereum-optimism/optimism/op-node/rollup"
-	"github.com/ethereum-optimism/optimism/op-node/rollup/sync"
-)
-
-func TestSyncTesterHFS_Canyon_CLSync(gt *testing.T) {
-	hardforks_ext.SyncTesterHFSExt(gt, rollup.Canyon, sync.CLSync)
-}
-
-func TestSyncTesterHFS_Canyon_ELSync(gt *testing.T) {
-	hardforks_ext.SyncTesterHFSExt(gt, rollup.Canyon, sync.ELSync)
-}
