--- op-acceptance-tests/tests/sync_tester/sync_tester_hfs_ext_isthmus/sync_tester_hfs_ext_isthmus_test.go
@@ -1,17 +1,0 @@
-package sync_tester_hfs_ext_isthmus
-
-import (
-	"testing"
-
-	"github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/sync_tester/hardforks_ext"
-	"github.com/ethereum-optimism/optimism/op-node/rollup"
-	"github.com/ethereum-optimism/optimism/op-node/rollup/sync"
-)
-
-func TestSyncTesterHFS_Isthmus_CLSync(gt *testing.T) {
-	hardforks_ext.SyncTesterHFSExt(gt, rollup.Isthmus, sync.CLSync)
-}
-
-func TestSyncTesterHFS_Isthmus_ELSync(gt *testing.T) {
-	hardforks_ext.SyncTesterHFSExt(gt, rollup.Isthmus, sync.ELSync)
-}
